LB Abgabe mit Duko und Feinschliffe, cya next time...

+++ todo/todo/src/main/java/wiss/todoApp/repository/RoleRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import wiss.todoApp.models.ERole;
-import wiss.todoApp.models.Role;
-
+import wiss.todoApp.models.*;
+// repoaitorys verbinden mit der db
+// oder spielen ine role dabei
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-    Optional<Role> findByName(ERole name);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+  Optional<Role> findByName(ERole name);
 }
